Refactor logic
