--- gatsbyjs/helper.ts
@@ -1,48 +1,0 @@
-import moment from 'moment';
-import siteConfig from '../data/SiteConfig';
-
-export const addSiblingNodes = (postNodes: any[]): any => (
-  createNodeField: any
-): void => {
-  postNodes.sort(
-    ({frontmatter: {date: date1}}, {frontmatter: {date: date2}}): number => {
-      const dateA = moment(date1, siteConfig.dateFromFormat);
-      const dateB = moment(date2, siteConfig.dateFromFormat);
-
-      if (dateA.isBefore(dateB)) return 1;
-
-      if (dateB.isBefore(dateA)) return -1;
-
-      return 0;
-    }
-  );
-
-  postNodes.forEach((_, i): void => {
-    const nextID = i + 1 < postNodes.length ? i + 1 : 0;
-    const prevID = i - 1 > 0 ? i - 1 : postNodes.length - 1;
-    const currNode = postNodes[i];
-    const nextNode = postNodes[nextID];
-    const prevNode = postNodes[prevID];
-
-    createNodeField({
-      node: currNode,
-      name: 'nextTitle',
-      value: nextNode.frontmatter.title,
-    });
-    createNodeField({
-      node: currNode,
-      name: 'nextSlug',
-      value: nextNode.fields.slug,
-    });
-    createNodeField({
-      node: currNode,
-      name: 'prevTitle',
-      value: prevNode.frontmatter.title,
-    });
-    createNodeField({
-      node: currNode,
-      name: 'prevSlug',
-      value: prevNode.fields.slug,
-    });
-  });
-};
