--- gatsbyjs/setFieldsOnGraphQLNodeType.ts
@@ -1,11 +1,0 @@
-export const setFieldsOnGraphQLNodeType = (addSiblingNodes: any): any => ({
-  type,
-  actions,
-}: any): void => {
-  const {name} = type;
-  const {createNodeField} = actions;
-
-  if (name === 'MarkdownRemark') {
-    addSiblingNodes(createNodeField);
-  }
-};
